separating type utils files
